mementobot/pyproject.toml
[tool.poetry]
name = "mementobot"
version = "0.1.0"
description = ""
authors = ["Lorenzo Carbonell <lorenzo.carbonell.cerezo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
dateparser = "^1.1.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
